refactor: address PR review feedback
- Extract MsalAccountHelper for DRY account lookup (Gemini feedback)
- Remove entityAlias from schema required array (Copilot feedback)
- Add CultureInfo.InvariantCulture to DateTimeOffset.Parse (Copilot feedback)

Generic catch clause deferred to #61.
